demos/pyproject.toml
[project]
name = "demos"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.58",
    "python-dotenv>=1.2.2",
]
